+++ src/main/java/hu/bmiklos/bc/service/UserService.java
@@ -1,6 +1,5 @@
 package hu.bmiklos.bc.service;
 
-import hu.bmiklos.bc.exception.NotAuthenticatedException;
 import hu.bmiklos.bc.exception.UserRegistrationException;
 import hu.bmiklos.bc.model.Email;
 import hu.bmiklos.bc.model.Password;
@@ -10,6 +9,8 @@ import hu.bmiklos.bc.repository.PasswordRepository;
 import hu.bmiklos.bc.repository.UserRepository;
 import hu.bmiklos.bc.service.dto.UserDto;
 import hu.bmiklos.bc.service.mapper.UserMapper;
+import hu.bmiklos.bc.business.security.ActiveUserService;
+import hu.bmiklos.bc.business.security.NotAuthenticatedException;
 import hu.bmiklos.bc.business.security.SaltedPasswordEncoder;
 import hu.bmiklos.bc.util.SaltGenerator;
 import jakarta.transaction.Transactional;
